# docs/openapi/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /admin/{id}/user:
    get:
      operationId: AdminController_getUser
      summary: Get a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
      tags:
        - admin
    patch:
      operationId: AdminController_updateUser
      summary: Update a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
      tags:
        - admin
  /admin/{id}/education:
    get:
      operationId: AdminController_getEducations
      summary: Get educations
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EducationDto'
      tags:
        - admin
    post:
      operationId: AdminController_createEducation
      summary: Creates a new education
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEducationDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationDto'
      tags:
        - admin
  /admin/{id}/education/{idEducation}:
    put:
      operationId: AdminController_updateEducation
      summary: Updates an education
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idEducation
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEducationDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationDto'
      tags:
        - admin
    delete:
      operationId: AdminController_deleteEducation
      summary: Deletes an education
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idEducation
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
  /admin/{id}/work-experience:
    get:
      operationId: AdminController_getWorkExperiences
      summary: Get work experiences
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkExperienceDto'
      tags:
        - admin
    post:
      operationId: AdminController_createWorkExperience
      summary: Creates a work experience
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkExperienceDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkExperienceDto'
      tags:
        - admin
  /admin/{id}/work-experience/{idWorkExperience}:
    put:
      operationId: AdminController_updateWorkExperience
      summary: Updates a work experience
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idWorkExperience
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkExperienceDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkExperienceDto'
      tags:
        - admin
    delete:
      operationId: AdminController_deleteWorkExperience
      summary: Deletes a work experience
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idWorkExperience
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
  /admin/{id}/project:
    get:
      operationId: AdminController_getProjects
      summary: Get projects
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
      tags:
        - admin
    post:
      operationId: AdminController_createProject
      summary: Creates a project
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
      tags:
        - admin
  /admin/{id}/project/{idProject}:
    put:
      operationId: AdminController_updateProject
      summary: Updates a project
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idProject
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
      tags:
        - admin
    delete:
      operationId: AdminController_deleteProject
      summary: Deletes a project
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idProject
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
  /admin/{id}/skill:
    get:
      operationId: AdminController_getSkills
      summary: Get skills
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillDto'
      tags:
        - admin
    post:
      operationId: AdminController_createSkill
      summary: Creates a skill
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSkillDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDto'
      tags:
        - admin
  /admin/{id}/skill/{idSkill}:
    put:
      operationId: AdminController_updateSkill
      summary: Updates a skill
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idSkill
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSkillDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDto'
      tags:
        - admin
    delete:
      operationId: AdminController_deleteSkill
      summary: Deletes a skill
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idSkill
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
  /admin/{id}/contact:
    get:
      operationId: AdminController_getContact
      summary: Get contacts
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactDto'
      tags:
        - admin
    post:
      operationId: AdminController_createContact
      summary: Creates a contact
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactDto'
      tags:
        - admin
  /admin/{id}/contact/{idContact}:
    put:
      operationId: AdminController_updateContact
      summary: Updates a contact
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idContact
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactDto'
      tags:
        - admin
    delete:
      operationId: AdminController_deleteContact
      summary: Deletes a contact
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idContact
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
...
  /admin/{id}/footer/{idFooter}:
    delete:
      operationId: AdminController_deleteFooter
      summary: Deletes a footer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idFooter
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - admin
info:
  title: Portfolio API
  description: The OpenAPI of the Portoflio API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          format: date-time
          type: string
        description:
          type: string
      required:
        - id
        - username
        - firstName
        - lastName
        - dateOfBirth
        - description
    UpdateUserDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          format: date-time
          type: string
        description:
          type: string
      required:
        - firstName
        - lastName
        - dateOfBirth
        - description
    EducationDto:
      type: object
      properties:
        id:
          type: string
        courseName:
          type: string
        institute:
          type: string
        instituteLink:
          type: string
        startDate:
          format: date-time
          type: string
        endDate:
          format: date-time
          type: string
        grade:
          type: number
        user:
          type: string
      required:
        - id
        - courseName
        - institute
        - instituteLink
        - startDate
        - endDate
        - grade
        - user
...
    CreateWorkExperienceDto:
      type: object
      properties:
        role:
          type: string
        company:
          type: string
        companyLink:
          type: string
        startDate:
          format: date-time
          type: string
        endDate:
          format: date-time
          type: string
      required:
        - role
        - company
        - companyLink
        - startDate
        - endDate
...